pyproject.toml
[tool.poetry]
name = "sphiwe"
version = "0.1.0"
description = ""
authors = ["Nazire Mathe <nazireemathe@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
django = "4.2.11"
django-split-settings = "^1.3.1"
dotmap = "^1.3.30"
pillow = "^10.3.0"
whitenoise = "^6.6.0"
requests = "^2.32.3"
django-crispy-forms = "^2.2"
crispy-bootstrap4 = "^2024.1"
django-axes = "^6.5.1"
django-otp = "^1.5.4"
qrcode = "^7.4.2"
django-recaptcha = "^4.0.0"
django-storages = "^1.14.4"
google-cloud-storage = "^2.18.2"
psycopg2-binary = "^2.9.9"
botocore = "^1.35.17"
boto3 = "^1.35.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
